requirements.txt
opencv-python>=4.8
numpy>=1.24
PyYAML>=6.0
mediapipe>=0.10
psutil>=5.9
